Rename util pkg to split to fix lint error
