+++ pkg/components/split/splititerator.go
