+++ pkg/components/split/splititerator_test.go
